refactor(fathers page): move Background component into its folder
